Calculate in 64-bit, output in 32-bit

+++ improver/ensemble_copula_coupling/ensemble_copula_coupling.py
@@ -562,7 +562,7 @@ class ConvertProbabilitiesToPercentiles(BasePlugin):
         )
 
         forecast_at_percentiles = interpolate_multiple_rows_same_y(
-            percentiles_as_fractions, probabilities_for_cdf, threshold_points,
+            percentiles_as_fractions.astype(np.float64), probabilities_for_cdf.astype(np.float64), threshold_points.astype(np.float64),
         )
         forecast_at_percentiles = forecast_at_percentiles.transpose()
